CFY-7719 Disable CELERY_ACKS_LATE (#375)
This was enabled in CFY-4865 as an optimization, however it is
an incorrect optimization in case of unreliable networks.

With CELERY_ACKS_LATE=True, the task will be kept in the queue
until it is finished. This means that if the network connection is
interrupted during executing a task, the task raises an exception,
but is still left in the queue. When the network connection is
restored, the same task is still left there, and will be executed
again.

Cloudify has its own retrying semantics, and we must not use
acks_late for that as well, because that inevitably leads to
tasks being executed more than once.

Additionally, when CELERY_ACKS_LATE is disabled, task prefetch must
be limited. With prefetch=0, all tasks are consumed as soon as they
come (even if another task is executing at the same time). If the
tasks are acked early, prefetch=0 would mean that the tasks are
effectively buffered in the worker instead of in the queue, and would
be losts in case of a worker crash.
This means that when acks_late is disabled, prefetch must be set to 1
in order to not prefetch any additional tasks, so that none are lost
in a crash.