CFY-7791 Use celery results for task finish notifications (#377)
* Move deriving the task queue name to RemoteWorkflowTask

It makes no sense to do it on the workflow_context side;
it might well be done on the Task side, which results in less
passing things around.

* Implement the celery result poller

This is the main part: add a component which polls celery tasks for
completion

* Use the new celery results poller

Actually use the new component

* Events must not attempt to update a terminated task

In case the results mechanism updates the task before the
events mechanism does, the events handler must still be able to cope
with it already being updated.

* Don't create app if it's already passed in

* Actually, raise if the app is already provided

This only happens on programming errors

* Whoops!

* Also handle OperationRetry