CFY-7852 Derive queue attributes under try/except (#382)
This part might also throw a NonRecoverableError, so it must
be under the try/except which ends up handling a ignore_failure.
Without this, executions that can't contact the agent (eg. an
uninstall execution with install_method: init_script, for which
the agent wasn't installed yet) will fail, even with ignore_failure.